Tested with arduino library
Using the the arduino library dacWrite() function leads to fluctuation.
Testing 8 and 250 out of 255 for high and low on the dac pin results in between 0-97 on adc pin for low and 3959-4095 for adc pin for high. Big fluctuations for what should be constant.
Multimeter reported the same voltage for each high/low. No changes on dac apparently, must be the adc.